refactor + add header
