+++ app/_ui/Form.tsx
@@ -1,0 +1,14 @@
+const Form = () => {
+  return (
+    <form>
+      {/* radio gender */}
+      {/* select age group */}
+      {/* select upper exercise  */}
+      {/* select core exercise */}
+      {/* select cardio exercise */}
+      {/* reset/submit button */}
+    </form>
+  );
+};
+
+export default Form;
