+++ app/_ui/Header.tsx
@@ -1,0 +1,11 @@
+const Header = () => {
+  return (
+    <h1 className=" my-16 text-center text-5xl  font-semibold uppercase tracking-wider text-stone-200 text-shadow-sm sm:text-6xl">
+      Calculate your <br />
+      Air Force
+      <br /> PT Test Score
+    </h1>
+  );
+};
+
+export default Header;
