+++ app/page.tsx
@@ -1,7 +1,11 @@
-import Image from "next/image";
+import Form from "./_ui/Form";
+import Header from "./_ui/Header";
 
 export default function Home() {
   return (
-    <div className="min-h-screen bg-gradient-to-r from-sky-800 to-indigo-800 p-5"></div>
+    <div className="min-h-screen bg-gradient-to-r from-sky-800 to-indigo-800 p-5">
+      <Header />
+      <Form />
+    </div>
   );
 }
